pyproject.toml
[project]
name = "tkinter-can"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-can>=4.6.1",
]
